+++ src/lib/p0014.cpp
@@ -1,5 +1,6 @@
 #include "leetcode.hpp"
 #include <string>
+#include <vector>
 
 std::string leetcode::Solution0014::longestCommonPrefix(const std::vector<std::string>& strs) {
     if (strs.empty()) {
@@ -12,7 +13,7 @@ std::string leetcode::Solution0014::longestCommonPrefix(const std::vector<std::s
         if (c == '\0') {
             return prefix;
         }
-        for (int j = 1; j < strs.size(); j++) {
+        for (int j = 1; j < static_cast<int>(strs.size()); j++) {
             if (const char other_c = strs[j][i]; other_c == '\0' || other_c != c) {
                 return prefix;
             }
